added test_cmd_merge_gz.py and needed test files"

+++ vtam/CommandMerge.py
@@ -9,7 +9,7 @@ from vtam.utils.FileSampleInformation import FileSampleInformation
 from vtam.utils.RunnerVSearch import RunnerVSearch
 from vtam.utils.VTAMexception import VTAMexception
 from vtam.utils.Logger import Logger
-from utils.FileCompression import FileCompression
+from vtam.utils.FileCompression import FileCompression
 
 
 class CommandMerge(object):
@@ -133,18 +133,18 @@ class CommandMerge(object):
     
         for mergedfasta in fastainfo_df[['mergedfasta']].drop_duplicates().values:
             mergedfasta = mergedfasta[0]
-            fasta_merged_abspath = os.path.join(fastadir, mergedfasta)
-            mergedfasta_compressor = FileCompression(fasta_merged_abspath)
-            mergedfasta_c = mergedfasta_compressor.gzip_compression()
-            mergedfasta_compressor.delete_file()
-            _, relPath = os.path.split(mergedfasta_c)
-            fastainfo_df.loc[fastainfo_df['mergedfasta'] == mergedfasta, 'mergedfasta'] = relPath
-            
-
-            #fastq_info_df_i['mergedfasta'] = fasta_merged_basename
+            if mergedfasta.endswith('.gz'):
+                fasta_merged_abspath = os.path.join(fastadir, mergedfasta)
+                mergedfasta_compressor = FileCompression(fasta_merged_abspath)
+                mergedfasta_c = mergedfasta_compressor.gzip_compression()
+                mergedfasta_compressor.delete_file()
+                _, relPath = os.path.split(mergedfasta_c)
+                fastainfo_df.loc[fastainfo_df['mergedfasta'] == mergedfasta, 'mergedfasta'] = relPath
+            else: 
+                fastq_info_df_i['mergedfasta'] = fasta_merged_basename
 
         
         fastainfo_df.to_csv(fastainfo, sep="\t", header=True, index=False)
         # SummaryFileMerge(params_dic=vsearch_args_dic, stats_df=stats_df).write('summary.txt')
 
-CommandMerge.main(fastadir='../../fastaMixed', fastainfo='../../fastainfoMixed.tsv', fastqdir='../../merge/fastqMixed', fastqinfo='../../merge/fastqinfoMixed_mfzr.tsv')
+# CommandMerge.main(fastadir='../../fastaMixed', fastainfo='../../fastainfoMixed.tsv', fastqdir='../../merge/fastqMixed', fastqinfo='../../merge/fastqinfoMixed_mfzr.tsv')
